Fix bugs in forecast-only mode (GFSv16)

+++ ush/hpssarch_gen.sh
@@ -41,14 +41,16 @@ if [ $type = "gfs" ]; then
   rm -f gfsb.txt
   rm -f gfs_pgrb2b.txt
   rm -f gfs_flux.txt
-  rm -f gfs_${format}a.txt
+  if [ $MODE = "cycled" ]; then
+    rm -f gfs_${format}a.txt
+    touch gfs_${format}a.txt
+  fi
   rm -f gfs_${format}b.txt
   rm -f gfs_restarta.txt
   touch gfsa.txt
   touch gfsb.txt
   touch gfs_pgrb2b.txt
   touch gfs_flux.txt
-  touch gfs_${format}a.txt
   touch gfs_${format}b.txt
   touch gfs_restarta.txt
 
@@ -63,23 +65,29 @@ if [ $type = "gfs" ]; then
   head="gfs.t${cyc}z."
 
   #..................
-  echo  "${dirname}${head}pgrb2b.0p25.anl                  " >>gfs_pgrb2b.txt
-  echo  "${dirname}${head}pgrb2b.0p25.anl.idx              " >>gfs_pgrb2b.txt
-  echo  "${dirname}${head}pgrb2b.0p50.anl                  " >>gfs_pgrb2b.txt
-  echo  "${dirname}${head}pgrb2b.0p50.anl.idx              " >>gfs_pgrb2b.txt
+  if [ $MODE = "cycled" ]; then
+    echo  "${dirname}${head}pgrb2b.0p25.anl                  " >>gfs_pgrb2b.txt
+    echo  "${dirname}${head}pgrb2b.0p25.anl.idx              " >>gfs_pgrb2b.txt
+    echo  "${dirname}${head}pgrb2b.0p50.anl                  " >>gfs_pgrb2b.txt
+    echo  "${dirname}${head}pgrb2b.0p50.anl.idx              " >>gfs_pgrb2b.txt
+  fi
 
   echo  "./logs/${CDATE}/gfs*.log                          " >>gfsa.txt
-  echo  "${dirname}${head}gsistat                          " >>gfsa.txt
-  echo  "${dirname}${head}nsstbufr                         " >>gfsa.txt
-  echo  "${dirname}${head}prepbufr                         " >>gfsa.txt
-  echo  "${dirname}${head}prepbufr_pre-qc                  " >>gfsa.txt
-  echo  "${dirname}${head}prepbufr.acft_profiles           " >>gfsa.txt
-  echo  "${dirname}${head}pgrb2.0p25.anl                   " >>gfsa.txt
-  echo  "${dirname}${head}pgrb2.0p25.anl.idx               " >>gfsa.txt
-  echo  "${dirname}avno.t${cyc}z.cyclone.trackatcfunix     " >>gfsa.txt
-  echo  "${dirname}avnop.t${cyc}z.cyclone.trackatcfunix    " >>gfsa.txt
-  echo  "${dirname}trak.gfso.atcfunix.${PDY}${cyc}         " >>gfsa.txt
-  echo  "${dirname}trak.gfso.atcfunix.altg.${PDY}${cyc}    " >>gfsa.txt
+  if [ $MODE = "cycled" ]; then
+    echo  "${dirname}${head}gsistat                          " >>gfsa.txt
+    echo  "${dirname}${head}nsstbufr                         " >>gfsa.txt
+    echo  "${dirname}${head}prepbufr                         " >>gfsa.txt
+    echo  "${dirname}${head}prepbufr_pre-qc                  " >>gfsa.txt
+    echo  "${dirname}${head}prepbufr.acft_profiles           " >>gfsa.txt
+    echo  "${dirname}${head}pgrb2.0p25.anl                   " >>gfsa.txt
+    echo  "${dirname}${head}pgrb2.0p25.anl.idx               " >>gfsa.txt
+  fi
+  if [ $VRFYTRAK = "YES" ]; then
+    echo  "${dirname}avno.t${cyc}z.cyclone.trackatcfunix     " >>gfsa.txt
+    echo  "${dirname}avnop.t${cyc}z.cyclone.trackatcfunix    " >>gfsa.txt
+    echo  "${dirname}trak.gfso.atcfunix.${PDY}${cyc}         " >>gfsa.txt
+    echo  "${dirname}trak.gfso.atcfunix.altg.${PDY}${cyc}    " >>gfsa.txt
+  fi
   echo  "${dirname}storms.gfso.atcf_gen.${PDY}${cyc}       " >>gfsa.txt
   echo  "${dirname}storms.gfso.atcf_gen.altg.${PDY}${cyc}  " >>gfsa.txt
 
@@ -103,11 +111,12 @@ if [ $type = "gfs" ]; then
    fi
   fi
 
-  echo  "${dirname}${head}pgrb2.0p50.anl                   " >>gfsb.txt
-  echo  "${dirname}${head}pgrb2.0p50.anl.idx               " >>gfsb.txt
-  echo  "${dirname}${head}pgrb2.1p00.anl                   " >>gfsb.txt
-  echo  "${dirname}${head}pgrb2.1p00.anl.idx               " >>gfsb.txt
-
+  if [ $MODE = "cycled" ]; then
+    echo  "${dirname}${head}pgrb2.0p50.anl                   " >>gfsb.txt
+    echo  "${dirname}${head}pgrb2.0p50.anl.idx               " >>gfsb.txt
+    echo  "${dirname}${head}pgrb2.1p00.anl                   " >>gfsb.txt
+    echo  "${dirname}${head}pgrb2.1p00.anl.idx               " >>gfsb.txt
+  fi
 
   fh=0
   while [ $fh -le $FHMAX_GFS ]; do
@@ -145,11 +154,13 @@ if [ $type = "gfs" ]; then
 
 
   #..................
-  echo  "${dirname}${head}atmanl${SUFFIX}            " >>gfs_${format}a.txt
-  echo  "${dirname}${head}sfcanl${SUFFIX}            " >>gfs_${format}a.txt
-  echo  "${dirname}${head}atmi*.nc                   " >>gfs_${format}a.txt
-  echo  "${dirname}${head}dtfanl.nc                  " >>gfs_${format}a.txt
-  echo  "${dirname}${head}loginc.txt                 " >>gfs_${format}a.txt
+  if [ $MODE = "cycled" ]; then
+    echo  "${dirname}${head}atmanl${SUFFIX}            " >>gfs_${format}a.txt
+    echo  "${dirname}${head}sfcanl${SUFFIX}            " >>gfs_${format}a.txt
+    echo  "${dirname}${head}atmi*.nc                   " >>gfs_${format}a.txt
+    echo  "${dirname}${head}dtfanl.nc                  " >>gfs_${format}a.txt
+    echo  "${dirname}${head}loginc.txt                 " >>gfs_${format}a.txt
+  fi
 
   #..................
   if [ $OUTPUT_HISTORY = ".true." ]; then
@@ -163,12 +174,28 @@ if [ $type = "gfs" ]; then
   fi
 
   #..................
-  echo  "${dirname}RESTART/*0000.sfcanl_data.tile1.nc  " >>gfs_restarta.txt
-  echo  "${dirname}RESTART/*0000.sfcanl_data.tile2.nc  " >>gfs_restarta.txt
-  echo  "${dirname}RESTART/*0000.sfcanl_data.tile3.nc  " >>gfs_restarta.txt
-  echo  "${dirname}RESTART/*0000.sfcanl_data.tile4.nc  " >>gfs_restarta.txt
-  echo  "${dirname}RESTART/*0000.sfcanl_data.tile5.nc  " >>gfs_restarta.txt
-  echo  "${dirname}RESTART/*0000.sfcanl_data.tile6.nc  " >>gfs_restarta.txt
+  if [ $MODE = "cycled" ]; then
+    echo  "${dirname}RESTART/*0000.sfcanl_data.tile1.nc  " >>gfs_restarta.txt
+    echo  "${dirname}RESTART/*0000.sfcanl_data.tile2.nc  " >>gfs_restarta.txt
+    echo  "${dirname}RESTART/*0000.sfcanl_data.tile3.nc  " >>gfs_restarta.txt
+    echo  "${dirname}RESTART/*0000.sfcanl_data.tile4.nc  " >>gfs_restarta.txt
+    echo  "${dirname}RESTART/*0000.sfcanl_data.tile5.nc  " >>gfs_restarta.txt
+    echo  "${dirname}RESTART/*0000.sfcanl_data.tile6.nc  " >>gfs_restarta.txt
+  elif [ $MODE = "forecast-only" ]; then
+    echo  "${dirname}INPUT/gfs_ctrl.nc        " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/gfs_data.tile1.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/gfs_data.tile2.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/gfs_data.tile3.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/gfs_data.tile4.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/gfs_data.tile5.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/gfs_data.tile6.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/sfc_data.tile1.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/sfc_data.tile2.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/sfc_data.tile3.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/sfc_data.tile4.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/sfc_data.tile5.nc  " >>gfs_restarta.txt
+    echo  "${dirname}INPUT/sfc_data.tile6.nc  " >>gfs_restarta.txt
+  fi
 
   #..................
   if [ $DO_WAVE = "YES" ]; then
